Chore: moved filter-Modal filter into components
